Moving input types to their own package.
