+++ src/main/kotlin/hu/netsurf/erp/input/SupplierInput.kt
@@ -1,4 +1,4 @@
-package hu.netsurf.erp.model
+package hu.netsurf.erp.input
 
 data class SupplierInput(
     val name: String,
